move companion services in subpackage
